Moved code to helpers.
