+++ pyproject.toml
@@ -13,6 +13,11 @@ dependencies = [
     "scipy (>=1.17.0,<2.0.0)"
 ]
 
+[project.optional-dependencies]
+dev = [
+    "scipy-stubs"
+]
+
 
 [build-system]
 requires = ["poetry-core>=2.0.0,<3.0.0"]
