+++ requirements.txt
@@ -4,5 +4,6 @@ redis==5.0.4
 pydantic==2.7.1
 pydantic-settings==2.12.0
 scipy==1.13.1
+scipy-stubs
 eth-account==0.11.2
 PyJWT==2.8.0
